Reorganized ts functions and renamed modules
